#10206 add pos_horz parameter to addIcon method
Via Ticket #10206, address the need to specify a y/horizontal offset
when specifying an icon to CKEditor.skins.addIcon.  That allows using
existing corporate colateral image sprites that may not be laid out in a
single column. when defining a new skin or plugin.

Adding the offset_horz after the offset_vert breaks the normal X before
Y paradigm, but doing it this way maintains backwards compability with
all existing source for plugins/skins.

Sorry for the previous commits/reverts.  Still new to github.