#10208 add override property to addIcon method
This change adds an additional "overrride" property to
CKEDITOR.skins.addIcon() [core/skins.js]

The property would be use to override the current check to see if an
existing object is stored in the CKEDITOR.skins.icons array, prior to
allowing a change to be made.

The use case for this is to enable "dropin non-built skins".  Skins that
are initialized at "run-time" via the config.js or via config:
initialization property:  config ex:
jQuery("#id").ckeditor({skin:'myskin'});

The "skins" property causes ckeditor to load the skins/myskin/skin.js
after ckeditor.js is initailized, thus allows such a skin to "override"
the values set by the skin that was used during the "build" of
ckeditor.js, via adding calls to addIcon in the skin.js file.