Ticket: #7280
The button panel needs to store its panel definition. Otherwise the panel definition is lost when switching toolbars which causes the text color button to no longer function afterwards. See comment comment 29 on the ticket for the code this was tested with. It is also part of attachment 7280_maximize.patch in the ticket, is seems no matter which manner of toolbar switching (integrated support or separate plugin) is used this patch part is needed.