Add editor.config.linkShowSurroundingAnchors
By default, editor.config.linkShowSurroundingAnchors is true,
and the 'Link to anchors in the text' option in the Link dialog
shows all anchors in the document in which the editor is rendered.

When editor.config.linkShowSurroundingAnchors is false, only
the anchors within the editable area are shown. This is especially
useful when the divarea plugin is used and not all anchors in the
surrounding page should show up as options in the Link dialog.