Fix tableresize to only attach to tables within containing editor
tableresize plugin attached mouse move handler to document body
element.  As a result any tables on the same page as a ckeditor that
was created using either divarea or inline editing will become
resizeable.

When an iframe is used, it is acceptable to attach to body.  When
inline or divarea is used, one must filter table resize events to
only those contained within an actual ckeditor.