Fix to ensure any ascendant tables found are within editor
When tableresize plugin attempts to determine if mouse is currently
over a table, it perform a search of the ascendants of the event
target.  However this search will continue outside of the editable
container, than thus find tables in which the editable may be
embedded.  This fix ads a check to ensure any found table is in
fact a descendant of the editable container.