save plugin inline editor ajax save optional feat.
Leaves the current functionality for `CKEDITOR.ELEMENT_MODE_REPLACE` without any modifications, adding a few extra lines for handling `CKEDITOR.ELEMENT_MODE_INLINE` editors: 

If a target is specified on an inline editor container's html5 `data-actionUrl` param and the AJAX plugin is available, the standard save button will show up in the editor. 

The button will show up disabled and it will be enabled only when the editor content is dirty. 

When clicked, it will trigger an AJAX POST request containing the editor id and content (post params `editorID` and `editabledata`, urlencoded).