Moved initial selection fix from the constructor to editable.
The reason why we must move this fix is that there’s no difference between initialising editor and setting data - in both cases the same element is active and anchor is the same. However, we need to force focusing editable in the second case and we cannot do that in the first case.

In editable we can check whether the editable has focus and the code will also be executed only once - not on every getSelection(), so we should have a small performance boost.