Fixing #13703 by checking adjacency before preventing deletion.
Most of the code here is a workaround for a long-standing firefox bug by reproducing what Chrome/IE do... merging the block tags.

There is a notable case that is still not fixed: If a widget is wrapped in a non-block tag, the widget will not be selected, it will be deleted.
This changeset does not attempt to fix that issue.