Update plugin.js
Per https://docs.cksource.com/ckeditor_api/symbols/CKEDITOR.ui.dialog.uiElement.html#event:change, whenever there is a change to the ui, the event 'change' should be fired.  This is particularly useful for Angular applications, where the event is required after the file upload in order for Angular to know that a change to the text has been made.