Updates to the image plugin to improve handling of percent-sized images.
Reworked the onSizeChange handler in image.js so that if percentages are entered, the lock ratio setting is enabled, not disabled, with the percentage value simply copied from width to height or vice versa.
Updated the commit function for the txtHeight control to compute a meaningful height for the preview when a percentage width is in use.
Updated the commit function for the txtHeight control to omit the height CSS attribute altogether if a percentage value is supplied, since a percentage height is meaningless for documents without a fixed height.
Updated setupDimension to fill an empty height with the width if it is a percentage, and to turn on lock ratio.
Updated checkDimension within setupDimension to turn the lock ratio checkbox on instead of off when a percentage value is detected.
Updated the 'check' path within switchLockRatio to detect percentage width/height values.