Giving a try to ContextManager class - basic interface so far.
It became clear that contexts will need some sort of manager in order to implement prioritized checks. Currently checking logic is implemented in the Context.refresh method - but the problem is that it makes all three checks at once.

Instead we need to do only highest priority checks (options.refresh) on each active context. If none matched, then we move on to second priority matching, which is widget - and so on.