Tests: removed unused inline toolbar context method.
Notes: Actually I had to remove it as this test didn't have sense anymore.

It assumed that options.widget will take effect if the selection is within nested editable, say a strong (and this case is covered with unit test in widget.js).

However after research we have implemented it in a way that options.widget takes effect only when the widget itself is focused.