Explicitly marked widget name in it's definition object.
This is done as a temp workaround, as the objects that we're passing as definition to features, are copies of the original definition. In other words any modifications to definition done at runtime (by other features setUps or at a widget runtime) are not available in setUp function. Because of that definition.name was not available (because it's assigned by the widget system at a widget construction time).