Improved IE 10 and IE 11 version detection
In IE 10 and IE 11, there are cases observed where document.documentMode may be undefined and the user agent does not match what CKEditor expects for detection.  For example, in IE11, the user agent may be "Mozilla/5.0 (Windows NT 6.3); Trident/7.0; rv:11.0) like Gecko) and documentMode can be undefined.  In this case, CKEditor will error out on line 196 with:

"SCRIPT5007: Unable to get property '1' of undefined or null reference"

The code added here does a slower but more rigorous check of the user agent.  For IE10 or IE11, if documentMode is undefined, the version is extracted from the user agent and set properly.  This makes it possible for CKEditor 4.10 to support IE11 (at least in the cases that we tested).